push: options

+++ push.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"errors"
 	"fmt"
 	"io"
 	"os"
@@ -10,10 +11,17 @@ import (
 func runPush(r *Repository, args []string, w io.Writer) error {
 	f := NewFlagSet("push", "[path ...]")
 	dryRun := f.Bool("n", false, "dry run")
-	_ = dryRun
 
 	if err := f.Parse(args); err != nil {
 		return err
+	}
+	where := make(map[string]struct{})
+	for _, arg := range f.Args() {
+		dest, err := filepath.Abs(arg)
+		if err != nil {
+			return err
+		}
+		where[dest] = struct{}{}
 	}
 	dir := filepath.Join(r.StateDir, "store")
 	err := filepath.WalkDir(dir, func(p string, d os.DirEntry, err error) error {
@@ -28,11 +36,21 @@ func runPush(r *Repository, args []string, w io.Writer) error {
 		if err != nil {
 			return err
 		}
+		if _, ok := where[state.Target]; !ok && len(where) > 0 {
+			return nil
+		}
+		defer func() {
+			delete(where, state.Target)
+		}()
 		h, err := ReadHash(state.Target)
 		if err != nil {
 			return err
 		}
 		if h != state.Hash {
+			if *dryRun {
+				fmt.Printf("cp %q %q\n", state.Target, state.Source)
+				return nil
+			}
 			h, mode, err := CopyFile(state.Source, state.Target, CopyFileOptions{
 				Overwrite: true,
 			})
@@ -47,10 +65,17 @@ func runPush(r *Repository, args []string, w io.Writer) error {
 			return err
 		}
 		if !ok {
+			if *dryRun {
+				fmt.Printf("chmod %o %q\n", mode, state.Source)
+				return nil
+			}
 			if err := os.Chmod(state.Source, mode); err != nil {
 				return err
 			}
 			state.Mode = mode
+		}
+		if *dryRun {
+			return nil
 		}
 		s := fmt.Sprintf("%s %o %s\n", state.Hash, state.Mode, state.Target)
 		return writeFile(p, []byte(s), FileOptions{})
@@ -60,6 +85,13 @@ func runPush(r *Repository, args []string, w io.Writer) error {
 			return nil
 		}
 		return err
+	}
+	if len(where) > 0 {
+		errs := make([]error, len(where))
+		for s := range where {
+			errs = append(errs, fmt.Errorf("'%s' does not changed", s))
+		}
+		return errors.Join(errs...)
 	}
 	return nil
 }
